Pequenas correções e otimizações.

+++ BibliotecaCentral/WebService/AutorizarNota/CorpoRequest.cs
@@ -7,12 +7,12 @@ namespace BibliotecaCentral.WebService.AutorizarNota
     {
         [XmlAttribute]
         public string versao { get; set; }
-        public int idLote { get; set; }
+        public long idLote { get; set; }
         public int indSinc { get; set; }
         [XmlElement(ElementName = nameof(NFe), Namespace = "http://www.portalfiscal.inf.br/nfe", IsNullable = false)]
         public NFe[] NFe { get; set; }
 
-        public CorpoRequest(NFe[] xmls, int numeroPrimeiraNota)
+        public CorpoRequest(NFe[] xmls, long numeroPrimeiraNota)
         {
             versao = "3.10";
             idLote = numeroPrimeiraNota;
